Correcciones autoevaluacion

+++ app/Http/Controllers/FuentesSecundarias/DocumentoAutoevaluacionController.php
@@ -2,24 +2,25 @@
 
 namespace App\Http\Controllers\FuentesSecundarias;
 
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Models\IndicadorDocumental;
-use App\Models\Factor;
-use App\Models\TipoDocumento;
-use App\Models\Dependencia;
-use App\Models\Proceso;
-use App\Models\Caracteristica;
 use App\Http\Requests\DocumentosAutoevaluacionRequest;
-use Illuminate\Support\Facades\Storage;
 use App\Models\Archivo;
+use App\Models\Caracteristica;
+use App\Models\Dependencia;
 use App\Models\DocumentoAutoevaluacion;
+use App\Models\Factor;
+use App\Models\IndicadorDocumental;
+use App\Models\Proceso;
+use App\Models\TipoDocumento;
 use DataTables;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Http\Testing\File;
 
 
 class DocumentoAutoevaluacionController extends Controller
 {
-     /**
+    /**
      * Instantiate a new controller instance.
      *
      * @return void
@@ -41,6 +42,7 @@ class DocumentoAutoevaluacionController extends Controller
     {
         return view('autoevaluacion.FuentesSecundarias.DocumentosAutoevaluacion.index');
     }
+
     /**
      * Process datatables ajax request.
      *
@@ -49,31 +51,44 @@ class DocumentoAutoevaluacionController extends Controller
     public function data(Request $request)
     {
         if ($request->ajax() && $request->isMethod('GET')) {
-            
+
             $documentos_autoevaluacion = DocumentoAutoevaluacion::with('indicadorDocumental.caracteristica.factor')
-            ->with('archivo', 'tipoDocumento', 'dependencia')
-            ->get();
+                ->with('archivo')
+                ->with(['tipoDocumento' => function($query){
+                    return $query->select('PK_TDO_Id', 'TDO_Nombre');
+                }])
+                ->with(['dependencia' => function($query){
+                    return $query->select('PK_DPC_Id', 'DPC_Nombre');
+                }])
+                ->get();
 
             return DataTables::of($documentos_autoevaluacion)
-                ->addColumn('archivo', function ($documento_autoevaluacion) {
+                ->addColumn('file', function ($documento_autoevaluacion) {
 
 
                     if (!$documento_autoevaluacion->archivo) {
-                        
 
-                        return '<a class="btn btn-default btn-xs" href="' . $documento_autoevaluacion->DOA_Link .
-                            '"target="_blank" role="button">Link</a>';    
-                    }
-                    else{
-                        
-                        return '<a class="btn btn-default btn-xs" href="' .
+
+                        return '<a class="btn btn-success btn-xs" href="' . $documento_autoevaluacion->DOA_Link .
+                            '"target="_blank" role="button">Descargar</a>';
+                    } else {
+
+                        return '<a class="btn btn-success btn-xs" href="' .
                             $documento_autoevaluacion->archivo->ruta .
 
-                            '" target="_blank" role="button">' . $documento_autoevaluacion->archivo->ACV_Nombre . '</a>';
+                            '" target="_blank" role="button">Descargar</a>';
 
                     }
                 })
-                ->rawColumns(['archivo'])
+                ->addColumn('nombre', function ($documento_autoevaluacion) {
+                    if ($documento_autoevaluacion->archivo) {
+                        return $documento_autoevaluacion->archivo->ACV_Nombre;
+                    } 
+                    else {
+                        return 'Link';
+                    }
+                })
+                ->rawColumns(['file'])
                 ->removeColumn('created_at')
                 ->removeColumn('updated_at')
                 ->make(true);
@@ -91,27 +106,28 @@ class DocumentoAutoevaluacionController extends Controller
         $id_lineamiento = Proceso::findOrFail(session()->get('id_proceso'))->FK_PCS_Lineamiento;
 
         $factores = Factor::has('caracteristica.indicadores_documentales')
-        ->where('FK_FCT_Lineamiento', '=', $id_lineamiento)
-        ->where('FK_FCT_estado', '=', '1')
-        ->get()
-        ->pluck('FCT_Nombre', 'PK_FCT_Id');
-        $dependencias =  Dependencia::pluck('DPC_Nombre', 'PK_DPC_Id');
+            ->where('FK_FCT_Lineamiento', '=', $id_lineamiento)
+            ->where('FK_FCT_estado', '=', '1')
+            ->get()
+            ->pluck('FCT_Nombre', 'PK_FCT_Id');
+        $dependencias = Dependencia::pluck('DPC_Nombre', 'PK_DPC_Id');
         $tipo_documentos = TipoDocumento::pluck('TDO_Nombre', 'PK_TDO_Id');
 
-        return view('autoevaluacion.FuentesSecundarias.DocumentosAutoevaluacion.create', 
-        compact('factores', 'dependencias', 'tipo_documentos'));
+        return view('autoevaluacion.FuentesSecundarias.DocumentosAutoevaluacion.create',
+            compact('factores', 'dependencias', 'tipo_documentos'));
     }
+
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(DocumentosAutoevaluacionRequest $request)
     {
-        
 
-        if($request->hasFile('archivo')){
+
+        if ($request->hasFile('archivo')) {
             $archivo = $request->file('archivo');
             $nombre = pathinfo($archivo->getClientOriginalName(), PATHINFO_FILENAME);
             $extension = $archivo->getClientOriginalExtension();
@@ -123,20 +139,20 @@ class DocumentoAutoevaluacionController extends Controller
             $archivos->ruta = $url;
             $archivos->save();
 
-            $id_archivo = $archivos->PK_ACV_Id;   
+            $id_archivo = $archivos->PK_ACV_Id;
         }
 
         $documento_auto = new DocumentoAutoevaluacion();
-        $documento_auto->fill($request->only(['IDO_Nombre', 
-        'DOA_Numero', 
-        'DOA_Anio',
-        'DOA_Link',
-        'DOA_DescripcionGeneral',
-        'DOA_ContenidoEspecifico',
-        'DOA_ContenidoAdicional',
-        'DOA_Observaciones']));
+        $documento_auto->fill($request->only(['IDO_Nombre',
+            'DOA_Numero',
+            'DOA_Anio',
+            'DOA_Link',
+            'DOA_DescripcionGeneral',
+            'DOA_ContenidoEspecifico',
+            'DOA_ContenidoAdicional',
+            'DOA_Observaciones']));
 
-        $documento_auto->FK_DOA_Archivo = isset($id_archivo)?$id_archivo:null;
+        $documento_auto->FK_DOA_Archivo = isset($id_archivo) ? $id_archivo : null;
         $documento_auto->FK_DOA_IndicadorDocumental = $request->get('PK_IDO_Id');
         $documento_auto->FK_DOA_TipoDocumento = $request->get('PK_TDO_Id');
         $documento_auto->FK_DOA_Dependencia = $request->get('PK_DPC_Id');
@@ -144,15 +160,15 @@ class DocumentoAutoevaluacionController extends Controller
         $documento_auto->save();
 
         return response(['msg' => 'Indicador documental registrado correctamente.',
-        'title' => '¡Registro exitoso!'
-    ], 200) // 200 Status Code: Standard response for successful HTTP request
-          ->header('Content-Type', 'application/json');
+            'title' => '¡Registro exitoso!'
+        ], 200)// 200 Status Code: Standard response for successful HTTP request
+        ->header('Content-Type', 'application/json');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -162,7 +178,7 @@ class DocumentoAutoevaluacionController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
@@ -181,21 +197,22 @@ class DocumentoAutoevaluacionController extends Controller
             ->pluck('CRT_Nombre', 'PK_CRT_Id');
 
         $indicadores = IndicadorDocumental::where('FK_IDO_Caracteristica', '=', $documento->indicadorDocumental->FK_IDO_Caracteristica)
-        ->pluck('IDO_Nombre', 'PK_IDO_Id');
+            ->pluck('IDO_Nombre', 'PK_IDO_Id');
         $dependencias = Dependencia::pluck('DPC_Nombre', 'PK_DPC_Id');
         $tipo_documentos = TipoDocumento::pluck('TDO_Nombre', 'PK_TDO_Id');
-        
+        $size = $documento->archivo?filesize(public_path($documento->archivo->ruta)):null;
+
         return view(
             'autoevaluacion.FuentesSecundarias.DocumentosAutoevaluacion.edit',
-            compact('documento', 'factores', 'caracteristicas', 'indicadores', 'dependencias', 'tipo_documentos')
-            );
+            compact('documento', 'factores', 'caracteristicas', 'indicadores', 'dependencias', 'tipo_documentos', 'size')
+        );
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(DocumentosAutoevaluacionRequest $request, $id)
@@ -219,18 +236,17 @@ class DocumentoAutoevaluacionController extends Controller
                 $id_archivo = $archivos->PK_ACV_Id;
 
 
-            }
-            else{
+            } else {
                 $archivos = new Archivo();
                 $archivos->ACV_Nombre = $nombre;
                 $archivos->ACV_Extension = $extension;
                 $archivos->ruta = $url;
                 $archivos->save();
 
-                $id_archivo = $archivos->PK_ACV_Id; 
-                
+                $id_archivo = $archivos->PK_ACV_Id;
+
             }
-            
+
         }
         $documento->fill($request->only([
             'IDO_Nombre',
@@ -242,8 +258,11 @@ class DocumentoAutoevaluacionController extends Controller
             'DOA_ContenidoAdicional',
             'DOA_Observaciones'
         ]));
+        if (isset($id_archivo)){
+            $documento->FK_DOA_Archivo = $id_archivo;
+        }
 
-        $documento->FK_DOA_Archivo = isset($id_archivo) ? $id_archivo : null;
+        
         $documento->FK_DOA_IndicadorDocumental = $request->get('PK_IDO_Id');
         $documento->FK_DOA_TipoDocumento = $request->get('PK_TDO_Id');
         $documento->FK_DOA_Dependencia = $request->get('PK_DPC_Id');
@@ -251,38 +270,41 @@ class DocumentoAutoevaluacionController extends Controller
 
 
         return response(['msg' => 'El Indicador documental ha sido modificado exitosamente.',
-                'title' => 'Indicador Modificado!'
-            ], 200) // 200 Status Code: Standard response for successful HTTP request
-                ->header('Content-Type', 'application/json');
+            'title' => 'Indicador Modificado!'
+        ], 200)// 200 Status Code: Standard response for successful HTTP request
+        ->header('Content-Type', 'application/json');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
         $documento = DocumentoAutoevaluacion::findOrfail($id);
-        $ruta = str_replace('storage', 'public', $documento->archivo->ruta);
-        Storage::delete($ruta);
-        $documento->archivo()->delete();
-        
+        if ($documento->archivo) {
+            $ruta = str_replace('storage', 'public', $documento->archivo->ruta);
+            Storage::delete($ruta);
+            $documento->archivo()->delete();
+        }
+        $documento->delete();
+
 
         return response(['msg' => 'El documento ha sido eliminado exitosamente.',
-                'title' => 'Documento Eliminado!'
-            ], 200) // 200 Status Code: Standard response for successful HTTP request
-                ->header('Content-Type', 'application/json');
+            'title' => 'Documento Eliminado!'
+        ], 200)// 200 Status Code: Standard response for successful HTTP request
+        ->header('Content-Type', 'application/json');
     }
 
     public function obtenerCaracteristicas($id)
     {
         $caracteristicas = Caracteristica::has('indicadores_documentales')
-        ->where('FK_CRT_Factor', '=', $id)
-        ->where('FK_CRT_estado', '=', '1')
-        ->get()
-        ->pluck('CRT_Nombre', 'PK_CRT_Id');
+            ->where('FK_CRT_Factor', '=', $id)
+            ->where('FK_CRT_estado', '=', '1')
+            ->get()
+            ->pluck('CRT_Nombre', 'PK_CRT_Id');
         return json_encode($caracteristicas);
     }
 }
